First attempt at preserving zoom/pan state on chart update
This appears to work, but it produces screen flicker, because scale options are processed by Chart.update before the zoom plugin's `beforeUpdate` has a chance to run.

I don't see a way around that within Chart.js's API, so I'm instead going to experiment with adding functions to get the current zoom bounds so that I can manage zooming myself.